+++ src/Controller/AppController.php
@@ -24,6 +24,7 @@ namespace MeCms\Controller;
 
 use App\Controller\AppController as BaseController;
 use Cake\I18n\I18n;
+use Cake\Network\Exception\InternalErrorException;
 use MeCms\Core\Plugin;
 
 /**
@@ -37,8 +38,9 @@ class AppController extends BaseController {
 	protected function _checkLastSearch($query_id = NULL) {
         $interval = config('security.search_interval');
 		
-        if(empty($interval))
+        if(empty($interval)) {
             return TRUE;
+        }
 		
 		$query_id = empty($query_id) ? NULL : md5($query_id);
 		
@@ -46,11 +48,13 @@ class AppController extends BaseController {
 		
 		if(!empty($last_search)) {
 			//Checks if it's the same search
-			if(!empty($query_id) && !empty($last_search['id']) && $query_id === $last_search['id'])
+			if(!empty($query_id) && !empty($last_search['id']) && $query_id === $last_search['id']) {
 				return TRUE;
+            }
 			//Checks if the interval has not yet expired
-			elseif(($last_search['time'] + $interval) > time())
+			elseif(($last_search['time'] + $interval) > time()) {
 				return FALSE;
+            }
 		}
 		
 		$this->request->session()->write('last_search', ['id' => $query_id, 'time' => time()]);
@@ -70,16 +74,19 @@ class AppController extends BaseController {
 		$path = Plugin::path('MeCms', 'src'.DS.'Locale');
 		
 		if(empty($config) || $config === 'auto') {
-			if(is_readable($path.DS.substr($language, 0, 5).DS.'me_cms.po'))
+			if(is_readable($path.DS.substr($language, 0, 5).DS.'me_cms.po')) {
 				return substr($language, 0, 5);
-			elseif(is_readable($path.DS.substr($language, 0, 2).DS.'me_cms.po'))
+            }
+			elseif(is_readable($path.DS.substr($language, 0, 2).DS.'me_cms.po')) {
 				return substr($language, 0, 2);
+            }
 		}
 		elseif(!empty($config)) {
             $file = $path.DS.$config.DS.'me_cms.po';
             
-			if(!is_readable($file))
-				throw new \Cake\Network\Exception\InternalErrorException(__d('me_tools', 'File or directory {0} not readable', $file));
+			if(!is_readable($file)) {
+				throw new InternalErrorException(__d('me_tools', 'File or directory {0} not readable', $file));
+            }
 			
 			return $config;
 		}
@@ -97,20 +104,24 @@ class AppController extends BaseController {
 		//Checks if the file was successfully uploaded
 		if(isset($file['error']) && $file['error'] == UPLOAD_ERR_OK && is_uploaded_file($file['tmp_name'])) {
 			//Updated the target, adding the filename
-			if(!file_exists($target.DS.$file['name']))
+			if(!file_exists($target.DS.$file['name'])) {
 				$target .= DS.$file['name'];
+            }
 			//If the file already exists, adds the name of the temporary file to the filename
-			else
+			else {
 				$target .= DS.pathinfo($file['name'], PATHINFO_FILENAME).'_'.basename($file['tmp_name']).'.'.pathinfo($file['name'], PATHINFO_EXTENSION);
+            }
 
             $upload = move_uploaded_file($file['tmp_name'], $file['target'] = $target);
             
 			//Checks if the file was successfully moved to the target directory
-			if(!$upload)
+			if(!$upload) {
 				$error = __d('me_cms', 'The file was not successfully moved to the target directory');
+            }
 		}
-		else
+		else {
 			$error = __d('me_cms', 'The file was not successfully uploaded');
+        }
 		
 		$this->set(am(['error' => empty($error) ? FALSE : $error], compact('file')));
 		
@@ -126,23 +137,25 @@ class AppController extends BaseController {
 	 * @uses MeTools\Network\Request::isAction()
 	 * @uses _getLanguage()
 	 * @uses isOffline()
-	 * @uses setLanguage()
 	 */
 	public function beforeFilter(\Cake\Event\Event $event) {        
 		//Checks if the site is offline
-		if($this->isOffline())
+		if($this->isOffline()) {
 			return $this->redirect(['_name' => 'offline']);
+        }
 		
 		//Checks if the user's IP address is banned
-		if(!$this->request->isAction('ip_not_allowed', 'Systems') && $this->request->isBanned())
+		if(!$this->request->isAction('ip_not_allowed', 'Systems') && $this->request->isBanned()) {
 			return $this->redirect(['_name' => 'ip_not_allowed']);
+        }
 		
 		//Sets the user's language
 		I18n::locale($this->_getLanguage());
 		
-		//If the current request has no prefix, it authorizes the current action
-		if(!$this->request->param('prefix'))
+		//Authorizes the current action, if this is not an admin request
+		if(!$this->request->isAdmin()) {
 			$this->Auth->allow($this->request->action);
+        }
 		
 		//Sets the paginate limit and the maximum paginate limit
 		//See http://book.cakephp.org/3.0/en/controllers/components/pagination.html#limit-the-maximum-number-of-rows-that-can-be-fetched
@@ -160,11 +173,12 @@ class AppController extends BaseController {
 	 */
 	public function beforeRender(\Cake\Event\Event $event) {
 		//Ajax layout
-		if($this->request->is('ajax'))
+		if($this->request->is('ajax')) {
 			$this->viewBuilder()->layout('MeCms.ajax');
+        }
 		
 		//Uses a custom View class (`MeCms.AppView` or `MeCms.AdminView`)
-		$this->viewClass = !$this->request->isAdmin() ? 'MeCms.View/App' : 'MeCms.View/Admin';
+		$this->viewClass = $this->request->isAdmin() ? 'MeCms.View/Admin' : 'MeCms.View/App';
 		
 		//Sets auth data for views
 		$this->set('auth', empty($this->Auth) ? FALSE : $this->Auth->user());
@@ -183,20 +197,21 @@ class AppController extends BaseController {
         $this->loadComponent('MeTools.Flash');
         $this->loadComponent('RequestHandler');
 		
-		if(config('security.recaptcha'))
+		if(config('security.recaptcha')) {
 			$this->loadComponent('MeTools.Recaptcha');
+        }
 		
 		parent::initialize();
     }
 	
 	/**
-	 * Checks if the provided user is authorized for the request
+	 * Checks if the user is authorized for the request
 	 * @param array $user The user to check the authorization of. If empty the user in the session will be used
 	 * @return bool TRUE if the user is authorized, otherwise FALSE
 	 * @uses MeCms\Controller\Component\AuthComponent::isGroup()
 	 */
 	public function isAuthorized($user = NULL) {		
-		//By default, admins and managers can access every action
+		//By default, admins and managers can access all actions
 		return $this->Auth->isGroup(['admin', 'manager']);
 	}
 	
@@ -206,16 +221,19 @@ class AppController extends BaseController {
 	 * @uses MeTools\Network\Request::isAction()
 	 */
 	protected function isOffline() {
-		if(!config('frontend.offline'))
+		if(!config('frontend.offline')) {
 			return FALSE;
-		
-		//Always online for these actions
-		if($this->request->isAction(['offline', 'login', 'logout']))
-			return FALSE;
+        }
 		
 		//Always online for admin requests
-		if($this->request->isAdmin())
+		if($this->request->isAdmin()) {
 			return FALSE;
+        }
+		
+		//Always online for these actions
+		if($this->request->isAction(['offline', 'login', 'logout'])) {
+			return FALSE;
+        }
 		
 		return TRUE;
 	}