+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 2.x branch
 ## 2.6 branch
+### 2.6.2
+* improved the code to check the cache validity. Removed `checkIfCacheIsValid()` and `getNextToBePublished()` methods;
+* improved code for posts and pages that are drafts or to be published in the future;
+* fixed bug on MenuBuilder helper;
+* updated Facebook's tags.
+
 ### 2.6.1
 * added userbar for frontend. It allows to edit an delete posts, pages and photos.